requirements.txt
absl-py==1.4.0
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
awsebcli==3.20.10
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blessed==1.20.0
botocore==1.31.85
cachetools==5.3.0
cement==2.8.2
certifi==2024.7.4
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.3
comm==0.1.2
contourpy==1.0.7
cycler==0.10.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==1.2.0
fastjsonschema==2.16.3
flatbuffers==23.3.3
fonttools==4.39.0
fqdn==1.5.1
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httpcore==1.0.2
httpx==0.25.2
idna==3.7
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.12.0rc1
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.9.1.0
mistune==2.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
openai==1.35.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opt-einsum==3.3.0
package_name==0.1
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.14.0
PyJWT==2.9.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.0.3
rsa==4.9
scipy==1.10.1
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.5.0
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0rc0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.19
wcwidth==0.1.9
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.5
wrapt==1.14.1